requirements.txt
streamlit
langgraph
langchain
langchain-groq
langchain-community
sympy
scipy
matplotlib
pandas
python-dotenv